requirements.txt
fastapi
uvicorn
pymongo
pinecone-client
sentence-transformers
google-generativeai
transformers
torch
langchain
langchain-community
tiktoken
python-dotenv
pypdf
pandas
numpy